+++ organisations/tests.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from .models import Organisation,Level,Team
+from .models import Organisation,Level,Team,Position
 from .seeder import gen_organisastion
 from qetela.schema import schema
 
@@ -24,6 +24,13 @@ def initialize():
         level=level
     )
     team.save()
+    position = Position(
+        id=1,
+        title='Test',
+        description='Test description',
+        team = team
+    )
+    position.save()
 
 class OrganisationGraphTests(TestCase):
     def test_organisations_with_no_records(self):
@@ -338,6 +345,107 @@ class TeamGraphTest(TestCase):
         '''
         expected = {
            "deleteTeam": {
+             "ok": True
+           }
+        }
+        result = schema.execute(mutation)
+        assert not result.errors
+        assert result.data == expected
+
+class PositionGraphTest(TestCase):
+    def test_get_organisation_with_position(self):
+        initialize()
+        query = '''
+        query organisation{
+          organisation(id:1){
+            name
+            levelSet{
+              label
+              teamSet{
+                name
+                positionSet{
+                  id,
+                  title
+                }
+              }
+            }
+          }
+        }
+        '''
+        expected = {
+           "organisation": {
+              "name": "Test",
+              "levelSet": [
+                {
+                  "label": "Test",
+                  "teamSet": [
+                    {
+                      "name": "Test",
+                      "positionSet": [
+                        {
+                          "id":"1",
+                          "title": "Test"
+                        }
+                       ]
+                    }
+                  ]
+                }
+                ]
+            }
+         }
+        result = schema.execute(query)
+        assert not result.errors
+        assert result.data == expected
+
+    def test_adding_position(self):
+        initialize()
+        mutation = '''
+        mutation addPosition{
+          addPosition(teamId:1,title:"Test",description:"Test description"){
+            title
+          }
+        }
+        '''
+        expected = {
+           "addPosition": {
+             "title": "Test"
+           }
+        }
+        result = schema.execute(mutation)
+        assert not result.errors
+        assert result.data == expected
+
+    def test_updating_position(self):
+        initialize()
+        mutation = '''
+        mutation updatePosition{
+          updatePosition(id:1,title:"Updated",description:"Updated description"){
+            title,
+            description
+          }
+        }
+        '''
+        expected = {
+           "updatePosition": {
+             "title": "Updated",
+             "description":"Updated description"
+           }
+        }
+        result = schema.execute(mutation)
+        assert not result.errors
+        assert result.data == expected
+
+    def test_deleting_position(self):
+        initialize()
+        mutation = '''
+        mutation deletePosition{
+          deletePosition(id:1){
+            ok
+          }
+        }
+        '''
+        expected = {
+           "deletePosition": {
              "ok": True
            }
         }