+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Team model, seeds, queries, mutations and tests
  - Level model, seeds, queries, mutations and tests
+ - Position model, seeds, queries, mutations and tests
+
 
 ## 0.0.1 - 2019-02-18